+++ src/transaction.ts
@@ -20,7 +20,7 @@ export class Transaction {
   private readonly savepoints: Transaction[] = [];
   private readonly options: PartlyRequired<TransactionOptions, 'type' | 'isolationLevel' | 'readOnly'>;
   private readonly parent: Transaction | null;
-  private readonly id: string;
+  readonly id: string;
   private readonly name: string;
   private finished: 'commit' | undefined;
   private connection: Connection | undefined;
